improvement: learning type predicates in TS function signatures
I recently discovered functions whose return types are a _type
predicate_.

It's really useful to transform types (narrowing down types) when you
check the type is compatible with what you expect.

It adds a bit of clarity in the code by moving the type guards in
dedicated functions.

(more details of type predicates:
https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates)

_Note: the commit also makes use of a second discovered TS feature,
the Type intersection (`&`) to “extend” existing types with more
attributes and make them compatible for narrowing. ✨_