core(api): keep AxiosResponse to not loose http response details
The current response interceptor was keeping only the response body of
every http API calls. However it's not what we want. We want to be
able to check response bodies but also response status for instance.

In this commit such need is not very visible (because we don't read
any response status) but for later changes it will be good, especially
for error handling.

The good thing about this change is that it removes the `any` type I
had to use to “trick” axios. ☺️