validate: add a `--dry-run` flag to the `bump deploy` command
This is the equivalent to the old CLI `validate` command. It didn't
make much sense to have a completely separate command for it as it's
actually only doing a “preview” of an actual deploy by validating the
given definition file.

Note: There's one small subtility in this PR. The current /validations
API endpoint alwas for non “dry-run” actions with the
auto_create_documentation parameter. Thus keeping the 'auto create'
feature when calling `bump deploy --dry-run` didn't make sense. I have
thus made sure to avoid the possibility to auto create a documentation
with this flag.

Closes #2