core: parse raw file definition to send to the API
The Bump API expects a raw string containing the definition content in
plain string. Right now the CLI serializes the parsed file/URL to JSON
before sending it to Bump.

However we used to always send the raw file content to our API so we
continue to do so for the main api definition with this commit.

Note: Even if we send the raw main definition, please note that
references resolved by the CLI will now always be sent serialized in
JSON.