fix: resolve accented filesystem path correctly in $RefParser lib
The internal `$refs` Class of the json-schema-ref-parser library stores
paths in a URI encoded fashion. This is to prevent confusion between #
hash internal references (mentioned in the full json path) and #
characters in the filesystem path.

This commit makes sure to fetch the parsed values from the `$refs`
object from the `$refs.values()` function (instead of the
`$refs.get($ref)` function).

I got a bit confused with the internals of the lib as I expected to be
able to provide the returned values of `$Refs.paths()` to the
`$Refs.get()` function transparently. But this does not work for
filesystem paths containing special characters. The best solution is
thus to fetch the `$refs.values()` object which has decoded file path
as keys. I.e. exactly what we need.