diff: fix two files diff when previous version has no changes
This is a case that was missed in #141 when the first created version
has no changes compared to the currently deployed version on Bump.

However in the case of two files diff, we need to continue and try to
publish the second file to check if there's a diff to display.