diff: fix the returned value of the cmd
This bug was introduced in #163 by removing the return value of the
command execution.

This is a quickfix to make the diff work correctly in our
Github-Action even though we should make a much bigger refactoring to
avoid using oclif Commands programmatically in the Github Action.

Until we have more time to do the refactoring (started in
https://github.com/paulRbr/cli/tree/refactoring-extract-diff-lib)
let's fix it.

(Note: I wanted to add a test case for this but it is not obvious to
do as the @oclif/test library doesn't give the return value in the
testing context. I tried to suggest a change upstream -
https://github.com/oclif/test/pull/203 - but it includes an upgrade
from v1 to v2 of the testing library and has bigger implications that
we have no time to do neither right now...!)