fix: yet another bugfix for the diff feature the diff is empty
This bug happens currently when the diff is empty and if the first
file version has a change with the currently deployed documentation on
bump.

When a second file is provided to the CLI we should only check the
returned result of that second version, not the result of the first
version.