fix: send relative paths for external file system paths on Windows
This commit is a bugfix for Windows based filesystem when external
references are defined on disk. Indeed we had a wrong regexp matchin
only unix style filesystem paths.

The fix consists of parsing windows style filesystem paths too.

I've tested it on a VM and the fix seems to work fine.

Closes #361