fix: make sure to send posix-style paths to Bump servers
As a followup to #360, we need to send only posix style paths to our
servers as that's how we handle references internally.

The fix in #360 was useful to read filesystem on windows machine (for
external references) but was not enough to make sure references in
parent or child directory work correctly.

Completely fixes #361