fix: rename the 'publish' custom script to 'release'
As detailed in the NP (library that helps for the releases) readme,
the custom script name shouldn't be `publish` (because it's an
internal `npm` defined script). cf
https://github.com/sindresorhus/np#release-script

I always had the publishing script that got stuck right before pushing
the git tag and never understood why. Now that it's fixed I get it...

If only I had searched for this before 😅. Anyway no more manual git
push tag for the next releases, just use `npm run release` ☺️🎉.

I've updated the #18 issue which describes the release process.