fix: deploy from a URL
In 2.7.0 we introduced the “hub” deployment to deploy a whole
directory. However we introduced a constrait on the input file
parameter of the `deploy` command to be an object on the
filesystem (due to the usage of `fs.statSync` function).

This commit makes sure to not crash if we pass a URL string and want
to deploy from a URL (instead of a fs filepath).