deps: make sure to force the minimum version of AsyncAPI spec
This commit makes sure to about the minimal version of the
`@asyncapi/spec` package which was locked to a version supporting
asyncapi 2.6.0, but not changed in the package.json file.

With this commit we make sure to always support asyncapi 2.6.0 even
for already existing installation (or cli used as a lib - in the
github-action for example).