diff: increase timeout to 2min
We decided in the past to hard code a 30s timeout for a diff within
the CLI, but this is quite problematic especially if we have too much
traffic at the same time.

Until we have more power in our queuing system, I believe it's best to
allow a bit more time before failing.

This commit increases the timeout value by 3. Going from 30seconds to
2 minutes.