fix: align exported core Diff interface with previous 2.8 signature
This is a fix after the major upgrade of the CLI (in #587) where we
broke the interface of the exported core `Diff` class and made its
usage as a lib more complicated.

Let's align back to what it used to be, and even simplify the
interface by making the constructor argument optional.