diff: add a `--no-fail-on-breaking` flag
This commit adds a `--no-*` version of the `--fail-on-breaking` flag
on the diff command to be able to disable the default behavior within
CI enviroments (where CI=1 env var is set for instance). Indeed, in
those environment the flag was always forced to true (without being
well documented) until now.

About #623