Add flag --preview to Deploy commande
Allow user to deploy a new version  temporary,
deployed in the context of the API (logo, CSS, group by tag/path ...).

In this commit, we add a new tag `--preview`,
that can be added to `bump deploy` command.
Output is a temporary version, accessible by a link
visible in CLI response, and invisible in deployment list on Bump.sh workspace.