fix(overlay): apply update on ALL array items (including first index)
This bug was introduced in the refactor of the overlay core lib of
2.9.3.

Javascript default falsy values can be tricky (especially for `0`
which is falsy). In the case of applying an overlay on an array, the
first index (`0`) would never be modified due to this bug.

We now compare the index to the literal `undefined` instead of
checking its “falsyness”.