fix(overlay): fix removal of elements of an array
With the refactoring made on the overlay command in
[v2.9.3](https://github.com/bump-sh/cli/releases/tag/v2.9.3) we
introduced a bug when `remove`ing elements if the overlay action was
targeting an array.

This was due to the exploded json paths which contains indexes, and
that those index change each time we remove an element.

We've fixed the bug with @AurelDbs by iterating over the list of paths
in a reversed order (we start by the end of the array instead of the
beginning).