fix(overlay): don't output yaml anchors
Fixes #708

When outputing the overlayed result of the `bump overlay` command, we
shouldn't export yaml aliases (`*ref` used with `&ref`) as this
feature of YAML is not well supported by tools.

Also, it's not the job of the overlay command to guess whether to use
YAML anchors or not.

With this commit we make sure to output yaml without refs.