diff: support --overlay flag on diff command
This commit is both a refactoring (to move the logic to apply overlays
inside the `definition.extractDefinition` to DRY the code and simplify
extracting an overlayed definition) and a new feature.

The new feature is to allow the `--overlay` flag on the `bump diff`
command to be able to apply a list of overlays to the files before
creating the diff.

The feature was asked by @JakeSCahill (thanks for asking it makes a
lot of sense to have it for the diff command too)